start reworking based on json output format
Realized that the model output wouldn't simply be a json array because it would need to do chain-of-thought-style analysis first.
Also realized that the thing that would differ between model providers would most likely be the 'chain of thought'-like parts before the final result, not the 'parse final description of graph, validate that, and turn it into a jgrapht object' logic chunks
