+++ app/src/main/java/com/example/dungeontest/graph/AiRespJsonIngester.kt
@@ -1,53 +1,22 @@
 package com.example.dungeontest.graph
 
-import android.util.Log
-import com.squareup.moshi.FromJson
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.JsonDataException
 import com.squareup.moshi.JsonReader
-import com.squareup.moshi.JsonWriter
-import com.squareup.moshi.ToJson
 
-class AiRespRoomAdapter : JsonAdapter<List<AiRespRoom>>() {
-    private val tag = javaClass.simpleName
+abstract class AiRespJsonIngester: JsonAdapter<List<AiRespRoom>>() {
+
 
     private val badValMsgPrefix =
         "When parsing the json object for a Room in the dungeon map's json, the value for the key"
     private val missingKeyMsgPrefix =
         "The json object for a Room in the dungeon map's json was missing the key"
 
-    @FromJson
-    override fun fromJson(reader: JsonReader): List<AiRespRoom> {
-        val aiRoomDescriptions: MutableList<AiRespRoom> = mutableListOf();
-
-        if (!reader.hasNext()) {
-            Log.w(tag, "empty json string fed to ai response room adapter's parser")
-            return aiRoomDescriptions
-        }
-        val firstTokenType = reader.peek()
-        if (firstTokenType != JsonReader.Token.BEGIN_ARRAY) {
-            throw JsonDataException("ai response room adapter expects a json array at root of json document but instead encountered ${reader.readJsonValue()}")
-        }
-        reader.beginArray()
-        while (reader.hasNext()) {
-            val nextTokenType = reader.peek()
-            if (nextTokenType == JsonReader.Token.BEGIN_OBJECT) {
-                aiRoomDescriptions += parseSingleRoom(reader)
-            } else {
-                throw JsonDataException("expected beginning of a json object corresponding to a room in the map, but instead encountered: ${reader.readJsonValue()}")
-            }
-        }
-        reader.endArray()
-        return aiRoomDescriptions
-    }
-
-
     fun parseSingleRoom(reader: JsonReader): AiRespRoom {
         reader.beginObject()
         var id: Int? = null
         var label: String? = null
         var neighbors: MutableList<Int>? = null;
-
 
         var detailsAboutAlreadyParsedFields = ""
         while (reader.hasNext()) {
@@ -163,11 +132,6 @@ class AiRespRoomAdapter : JsonAdapter<List<AiRespRoom>>() {
         reader.endObject()
         //makes sure list in data object is immutable
         return AiRespRoom(id, label, neighbors.toList())
-    }
-
-    @ToJson
-    override fun toJson(writer: JsonWriter, value: List<AiRespRoom>?) {
-        TODO("Not implemented")//not relevant for project
     }
 
 }
